[#297] Refactor json-ld core literal datatypes
